[ping] 0.7.0
Complete rework. now shows with colours on each ping. example if the ping goes above 150 it will show the colour orange. and above 250 it'd show red.

Note that this only change colours on Discord WS. This no longer has non embed version.

If you want the colours to work on typing and messages, you're free to make an PR to make that work.