Well flag emojis will delete at 4 on limit 6.
I cannot handle that it wont, it seems to be only issue with flags so far. i will not care if it does, whats the point spamming it anyway?, right, whats the point... just let it delete at 4 if the limit is 6. raise the limit higher if you want to spam flags.